refactoring to use new Zendex HTTPoison Client
This change also replaces the tests that were using the "in-memory" client
with `exvcr` cassettes based tests. This does increase test time by about a
second, but the overall improvement in not having to maintain a separate
module for testing is nice. Mocking request/response pairs is, however,
slightly tedious.